requirements.txt
pillow
keyboard
pystray
google-generativeai
customtkinter
